+++ Main/Features/TroopWeight/Hooks/IOnCampaignUIHelperGetPartyHealthTooltip.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using TaleWorlds.CampaignSystem.Party;
+using TaleWorlds.Core.ViewModelCollection.Information;
+
+namespace TAOM.Features.TroopWeight.Hooks;
+
+public interface IOnCampaignUIHelperGetPartyHealthTooltip
+{
+    void OnGetPartyHealthTooltip(PartyBase party, ref List<TooltipProperty> __result);
+}