+++ Main/Features/TroopWeight/Hooks/IOnGameMenuPartyItemRefreshCounts.cs
@@ -1,0 +1,8 @@
+using TaleWorlds.CampaignSystem.ViewModelCollection.GameMenu.Overlay;
+
+namespace TAOM.Features.TroopWeight.Hooks;
+
+public interface IOnGameMenuPartyItemRefreshCounts
+{
+    void OnGameMenuPartyItemRefreshCounts(GameMenuPartyItemVM partyItem);
+}