+++ Main/Features/TroopWeight/Hooks/IOnPartyBaseHelperGetPartySizeText.cs
@@ -1,0 +1,9 @@
+using TaleWorlds.CampaignSystem.Party;
+using TaleWorlds.Localization;
+
+namespace TAOM.Features.TroopWeight.Hooks;
+
+public interface IOnPartyBaseHelperGetPartySizeText
+{
+    void OnGetPartySizeText(PartyBase party, ref TextObject __result);
+}